requirements.txt
pyautogen

# -----------
matplotlib 
numpy
